🗺️ Atlas: module coupling refactor
